+++ example-simple/src/main/java/com/zhuinden/simplestackcomposedogexample/MainActivity.kt
@@ -8,23 +8,23 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import com.zhuinden.simplestack.AsyncStateChanger
 import com.zhuinden.simplestack.History
-import com.zhuinden.simplestack.SimpleStateChanger
 import com.zhuinden.simplestack.navigator.Navigator
 import com.zhuinden.simplestackcomposeintegration.core.BackstackProvider
-import com.zhuinden.simplestackcomposeintegration.core.SimpleComposeStateChanger
+import com.zhuinden.simplestackcomposeintegration.core.ComposeStateChanger
 import com.zhuinden.simplestackextensions.navigatorktx.androidContentFrame
 import com.zhuinden.simplestackextensions.services.DefaultServiceProvider
 
 class MainActivity : AppCompatActivity() {
-    private val composeStateChanger = SimpleComposeStateChanger()
+    private val composeStateChanger = ComposeStateChanger()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         val backstack = Navigator.configure()
             .setScopedServices(DefaultServiceProvider())
-            .setStateChanger(SimpleStateChanger(composeStateChanger))
+            .setStateChanger(AsyncStateChanger(composeStateChanger))
             .install(this, androidContentFrame, History.of(FirstKey()))
 
         setContent {